Rename file in details.
