+++ packages/Search/src/details/CMakeLists.txt
@@ -18,7 +18,7 @@ APPEND_SET(HEADERS ${DIR}/${PACKAGE_NAME}_ETIHelperMacros.h)
 IF (${PACKAGE_NAME}_ENABLE_EXPLICIT_INSTANTIATION)
   # Generate ETI .cpp files for DataTransferKit::TreeConstruction.
   DTK_PROCESS_ALL_N_TEMPLATES(TREECONSTRUCTION_OUTPUT_FILES
-    "DTK_ETI_NT.tmpl" "TreeConstruction" "TREECONSTRUCTION"
+    "DTK_ETI_NT.tmpl" "DetailsTreeConstruction" "TREECONSTRUCTION"
     "${${PACKAGE_NAME}_ETI_NODES}" TRUE)
   LIST(APPEND SOURCES ${TREECONSTRUCTION_OUTPUT_FILES})
 ENDIF()
