+++ packages/Search/src/details/DTK_DetailsAlgorithms.hpp
@@ -9,9 +9,9 @@
 #ifndef DTK_DETAILS_ALGORITHMS_HPP
 #define DTK_DETAILS_ALGORITHMS_HPP
 
-#include <DTK_Box.hpp>
+#include <DTK_DetailsBox.hpp>
+#include <DTK_DetailsPoint.hpp>
 #include <DTK_KokkosHelpers.hpp>
-#include <DTK_Point.hpp>
 
 #include <Kokkos_Macros.hpp>
