+++ packages/Search/src/details/DTK_DetailsNode.hpp
@@ -10,7 +10,7 @@
 #ifndef DTK_NODE_HPP
 #define DTK_NODE_HPP
 
-#include <DTK_Box.hpp>
+#include <DTK_DetailsBox.hpp>
 #include <Kokkos_Pair.hpp>
 
 namespace DataTransferKit
