+++ packages/Search/src/details/DTK_DetailsTreeTraversal.hpp
@@ -12,13 +12,13 @@
 #include <DTK_DBC.hpp>
 
 #include <DTK_DetailsAlgorithms.hpp>
-#include <DTK_Predicate.hpp>
-#include <DTK_PriorityQueue.hpp>
-#include <DTK_Stack.hpp>
+#include <DTK_DetailsNode.hpp>
+#include <DTK_DetailsPredicate.hpp>
+#include <DTK_DetailsPriorityQueue.hpp>
+#include <DTK_DetailsStack.hpp>
 
 #include <DTK_BVHQuery.hpp>
 #include <DTK_LinearBVH.hpp>
-#include <DTK_Node.hpp>
 
 namespace DataTransferKit
 {
