+++ packages/Search/test/tstDetailsTreeTraversal.cpp
@@ -6,8 +6,8 @@
  * distributed under a BSD 3-clause license. For the licensing terms see    *
  * the LICENSE file in the top-level directory.                             *
  ****************************************************************************/
-#include <DTK_PriorityQueue.hpp>
-#include <DTK_Stack.hpp>
+#include <DTK_DetailsPriorityQueue.hpp>
+#include <DTK_DetailsStack.hpp>
 
 #include <Teuchos_UnitTestHarness.hpp>
